Fixes structure
