+++ Frontend/src/components/App.jsx
@@ -1,5 +1,6 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "resources/logo.svg";
+import "css/App.css";
+import Window from "components/Window";
 
 function App() {
   return (
@@ -9,6 +10,7 @@ function App() {
         <p>
           Edit <code>src/App.js</code> and save to reload.
         </p>
+        <Window></Window>
         <a
           className="App-link"
           href="https://reactjs.org"
