+++ Frontend/src/components/Window.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class Window extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>Hey I am a window component</h1>
+      </div>
+    );
+  }
+}
+
+export default Window;
